dvc: separate cache and remote classes
Finalizes code part of #4050
