+++ tests/func/test_data_cloud.py
@@ -6,11 +6,11 @@ import pytest
 from flaky.flaky_decorator import flaky
 
 from dvc.cache import NamedCache
+from dvc.cache.base import STATUS_DELETED, STATUS_NEW, STATUS_OK
 from dvc.external_repo import clean_repos
 from dvc.main import main
-from dvc.remote.base import STATUS_DELETED, STATUS_NEW, STATUS_OK
-from dvc.remote.local import LocalTree
 from dvc.stage.exceptions import StageNotFound
+from dvc.tree.local import LocalTree
 from dvc.utils.fs import remove
 from dvc.utils.yaml import dump_yaml, load_yaml
