+++ tests/func/test_import.py
@@ -234,7 +234,7 @@ def test_download_error_pulling_imported_stage(tmp_dir, dvc, erepo_dir):
     remove(dst_cache)
 
     with patch(
-        "dvc.remote.local.LocalTree._download", side_effect=Exception
+        "dvc.tree.local.LocalTree._download", side_effect=Exception
     ), pytest.raises(DownloadError):
         dvc.pull(["foo_imported.dvc"])
